Time: 5 ms (22.96%) | Memory: 8.4 MB (99.66%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool isPalindrome(int x) {
-        long reverse=0;
+        long long reverse=0;
         int remainder;
         int temp=x;
         while(temp>0)
